Moved Platform methods to Mixin
