+++ pepys_import/core/store/common_db.py
@@ -35,7 +35,7 @@ class SensorMixin:
         sensor_type = data_store.db_classes.SensorType().search_sensor_type(
             session, sensor_type
         )
-        host = data_store.db_classes.Platform().search_platform(session, host)
+        host = data_store.db_classes.Platform().search_platform(data_store, host)
 
         sensor_obj = data_store.db_classes.Sensor(
             name=name, sensor_type_id=sensor_type.sensor_type_id, host=host.platform_id,
@@ -44,3 +44,57 @@ class SensorMixin:
         session.flush()
 
         return sensor_obj
+
+
+class PlatformMixin:
+    @classmethod
+    def search_platform(cls, data_store, name):
+        # search for any platform with this name
+        Platform = data_store.db_classes.Platform
+        return data_store.session.query(Platform).filter(Platform.name == name).first()
+
+    def get_sensor(self, data_store, sensor_name=None, sensor_type=None, privacy=None):
+        """
+         Lookup or create a sensor of this name for this :class:`Platform`.
+         Specified sensor will be added to the :class:`Sensor` table.
+         It uses find_sensor method to search existing sensors.
+
+        :param data_store: DataStore object to to query DB and use missing data resolver
+        :type data_store: DataStore
+        :param sensor_name: Name of :class:`Sensor`
+        :type sensor_name: String
+        :param sensor_type: Type of :class:`Sensor`
+        :type sensor_type: SensorType
+        :param privacy: Privacy of :class:`Sensor`
+        :type privacy: Privacy
+        :return: Created :class:`Sensor` entity
+        :rtype: Sensor
+        """
+        Sensor = data_store.db_classes.Sensor
+
+        sensor = Sensor().find_sensor(data_store, sensor_name, self.platform_id)
+        if sensor:
+            return sensor
+
+        if sensor_type is None or privacy is None:
+            resolved_data = data_store.missing_data_resolver.resolve_sensor(
+                data_store, sensor_name, sensor_type, privacy
+            )
+            # It means that new sensor added as a synonym and existing sensor returned
+            if isinstance(resolved_data, Sensor):
+                return resolved_data
+            elif len(resolved_data) == 3:
+                (sensor_name, sensor_type, privacy,) = resolved_data
+
+        assert isinstance(
+            sensor_type, data_store.db_classes.SensorType
+        ), "Type error for Sensor Type entity"
+        # TODO: we don't use privacy for sensor. Is it necessary to resolve it?
+        # assert isinstance(privacy, Privacy), "Type error for Privacy entity"
+
+        return Sensor().add_to_sensors(
+            data_store=data_store,
+            name=sensor_name,
+            sensor_type=sensor_type.name,
+            host=self.name,
+        )
